requirements.txt
praw==5.4.0
python-chess>=0.23.9
cairosvg>=2.1.3
imageio>=2.3.0
giphypop>=0.3
numpy